pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diplomat-gate"
version = "0.2.0"
description = "Approval gates for AI agent payments and emails. CONTINUE / REVIEW / STOP before execution."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = []

authors = [
    { name = "Diplomat Services", email = "contact@diplomat.run" },
]

keywords = [
    "ai-agents", "tool-calls", "safety", "governance",
    "payments", "email", "guardrails", "approval-gate",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
rich = ["rich>=13.0"]
langchain = ["langchain-core>=0.1"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
all = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "langchain-core>=0.1",
    "openai>=1.0",
    "anthropic>=0.20",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pyyaml>=6.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://diplomat.run"
Repository = "https://github.com/Diplomat-ai/diplomat-gate"
Documentation = "https://github.com/Diplomat-ai/diplomat-gate#readme"
Issues = "https://github.com/Diplomat-ai/diplomat-gate/issues"

[project.scripts]
diplomat-gate = "diplomat_gate.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/diplomat_gate"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
branch = true
source = ["diplomat_gate"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
